feat: CRUDify event
